generate bill

+++ Model/MeterValue.swift
@@ -12,14 +12,14 @@ typealias MeterValueId = UUID
 struct MeterValue {
     let id: MeterValueId
     let date: Date
-    let value: Double
+    let value: Decimal
     let isPaid: Bool
 }
 
 extension MeterValue: Equatable, Hashable { }
 
 extension MeterValue {
-    static func initial(_ value: Double) -> MeterValue {
+    static func initial(_ value: Decimal) -> MeterValue {
         MeterValue(
             id: MeterValueId(),         // should be ignored!
             date: Date(),
